Misc. Changes
This is collection of small changes.  Most are very minor.
The one major change is adding a new class to handle submitting
a workflow to Cromwell.  This will replace the one in pysub.
This also moves the input and label parameters into the wf_configs.yaml.
This should make it more extensible.  The config file also adds to new
workflows that are used for reannotating or relabeling previously
processed data.